+++ src/Framework/StringParent.cpp
@@ -182,7 +182,7 @@ void StringParent::loadFromString(const wchar_t * s, wchar_t delim)
   {
     if (!*p || *p == delim)
     {
-      int len = __min((int)(p - pp), 4095);
+      size_t len = __min((size_t)(p - pp), 4095);
       wcsncpy_s(buf, 4096, pp, len);
       buf[len] = 0;
       pp = p + 1;
